[wip] further ClientSettings work
