+++ Projects/CoX/Common/GameData/clientoptions_definitions.h
@@ -1,28 +1,15 @@
-/*
- * Super Entity Game Server Project
- * https://github.com/Segs/Segs
- * Copyright (c) 2006 - 2018 Super Entity Game Server Team (see Authors.txt)
- * This software is licensed! (See License.txt for details)
- *
- */
-
 #pragma once
-#include "MapEvents.h"
+
+#include <QtCore/QString>
+#include <stdint.h>
+#include <vector>
+#include <cassert>
 
 enum WindowVisibility : uint32_t {
   wv_HideAlways     = 0,
   wv_Always         = 1,
   wv_OnMouseOver    = 2,
   wv_Selected       = 4,
-};
-
-class SaveClientOptions final : public MapLinkEvent
-{
-public:
-    ClientOptions data;
-    SaveClientOptions() : MapLinkEvent(MapEventTypes::evSaveClientOptions) {}
-    void serializeto(BitStream &bs) const override;
-    void serializefrom(BitStream &bs) override;
 };
 
 struct ClientOption
