+++ Projects/CoX/Servers/MapServer/MapEvents.h
@@ -586,7 +586,7 @@ public:
     }
 };
 
-//#include "Events/ClientOptions.h"
+#include "Events/SaveClientOptions.h"
 #include "Events/GameCommandList.h"
 #include "Events/ChatDividerMoved.h"
 #include "Events/SceneEvent.h"
